requirements.txt
mysql-connector-python==8.0.26
Flask-MySQLdb==0.2.0
flask-sqlalchemy>=2.5.1
mysql-connector>=2.2.9
python-dotenv==0.18.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
